refactor(egopulse): split tools.rs into submodules
